refactor endpoints into own classes for better encapsulation
