+++ ksql-rest-app/src/main/java/io/confluent/ksql/api/endpoints/PullQueryPublisher.java
@@ -13,11 +13,12 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package io.confluent.ksql.api.plugin;
+package io.confluent.ksql.api.endpoints;
 
 import io.confluent.ksql.GenericRow;
 import io.confluent.ksql.api.spi.QueryPublisher;
 import io.confluent.ksql.reactive.BufferedPublisher;
+import io.confluent.ksql.rest.entity.TableRowsEntity;
 import io.vertx.core.Context;
 import java.util.ArrayList;
 import java.util.List;
@@ -28,7 +29,7 @@ public class PullQueryPublisher extends BufferedPublisher<GenericRow> implements
   private final List<String> columnNames;
   private final List<String> columnTypes;
 
-  public PullQueryPublisher(final Context ctx, final TableRows tableRows,
+  public PullQueryPublisher(final Context ctx, final TableRowsEntity tableRows,
       final List<String> columnNames, final List<String> columnTypes) {
     super(ctx, toGenericRows(tableRows));
     this.columnNames = Objects.requireNonNull(columnNames);
@@ -36,7 +37,7 @@ public class PullQueryPublisher extends BufferedPublisher<GenericRow> implements
   }
 
   @SuppressWarnings({"unchecked", "rawtypes"})
-  private static List<GenericRow> toGenericRows(final TableRows tableRows) {
+  private static List<GenericRow> toGenericRows(final TableRowsEntity tableRows) {
     final List<GenericRow> genericRows = new ArrayList<>(tableRows.getRows().size());
     for (List row : tableRows.getRows()) {
       final GenericRow genericRow = GenericRow.fromList(row);
