--- ksql-rest-app/src/main/java/io/confluent/ksql/api/plugin/PullQueryApiExecutor.java
@@ -1,28 +1,0 @@
-/*
- * Copyright 2020 Confluent Inc.
- *
- * Licensed under the Confluent Community License (the "License"); you may not use
- * this file except in compliance with the License.  You may obtain a copy of the
- * License at
- *
- * http://www.confluent.io/confluent-community-license
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package io.confluent.ksql.api.plugin;
-
-import io.confluent.ksql.parser.tree.Query;
-import io.confluent.ksql.services.ServiceContext;
-import io.confluent.ksql.statement.ConfiguredStatement;
-
-public interface PullQueryApiExecutor {
-
-  TableRows execute(
-      ConfiguredStatement<Query> statement,
-      ServiceContext serviceContext
-  );
-}
